Fix CPU monitor issues.
Sometimes, the tick value doesn’t increase in between fetch
calls (maybe the system is blocked on I/O?).  This caused
divide-by-zero issues, which I added a `zerop` check around.  That
made the monitor return `nil`, which the caused the sparkline code to
choke.

New approach is to return the last stored value.  If ticks haven’t
changed, there’s nothing to recompute, so the monitor value ought to
be the same.